moved preloder includes, checked popupwindows
