+++ interface/.meteor/versions
@@ -33,7 +33,7 @@ ejson@1.0.12
 ethereum:accounts@0.3.10
 ethereum:blocks@0.3.1
 ethereum:dapp-styles@0.5.6
-ethereum:elements@0.7.7
+ethereum:elements@0.7.9
 ethereum:tools@0.5.2
 ethereum:web3@0.15.3
 fastclick@1.0.12
