+++ interface/client/styles/popupWindows.import.less
@@ -62,6 +62,7 @@
 		right: 0;
 		background: linear-gradient(transparent 0%, @colorGrayLight 20%);
 		padding: 0 @gridHeight;
+        padding-top: @gridHeight * 3;
 	}
 
 	.info.dapp-error{
@@ -86,7 +87,7 @@
 		.container {
 			overflow: auto;
 			height: 80vh;
-			padding-bottom: 160px;
+			padding-bottom: 120px;
 		}
 
 		.transaction-parties {
