+++ modules/preloader/browser.js
@@ -6,15 +6,13 @@ const electron = require('electron');
 const ipc = electron.ipcRenderer;
 const shell = electron.shell;
 const mist = require('../mistAPI.js');
-require('../openExternal.js');
 const BigNumber = require('bignumber.js');
 const ipcProviderWrapper = require('../ipc/ipcProviderWrapper.js');
 var Web3 = require('web3');
-require('../getFavicon.js');
-require('../getMetaTags.js');
-require('../openExternal.js');
+require('./include/getFavicon.js');
+require('./include/getMetaTags.js');
+require('./include/openExternal.js');
 require('./include/setBasePath')('interface');
-
 
 // notifiy the tab to store the webview id
 ipc.sendToHost('setWebviewId');
