+++ modules/preloader/dapps.js
@@ -9,9 +9,9 @@ const mist = require('../mistAPI.js');
 const BigNumber = require('bignumber.js');
 const ipcProviderWrapper = require('../ipc/ipcProviderWrapper.js');
 var Web3 = require('web3');
-require('../getFavicon.js');
-require('../getMetaTags.js');
-require('../openExternal.js');
+require('./include/getFavicon.js');
+require('./include/getMetaTags.js');
+require('./include/openExternal.js');
 require('./include/setBasePath')('interface');
 
 // notifiy the tab to store the webview id
@@ -19,7 +19,6 @@ ipc.sendToHost('setWebviewId');
 
 // destroy the old socket
 ipc.send('ipcProvider-destroy');
-
 
 window.mist = mist();
 window.BigNumber = BigNumber;
