+++ modules/preloader/popupWindows.js
@@ -6,16 +6,16 @@ require('./include/common')('popup');
 const electron = require('electron');
 const ipc = electron.ipcRenderer;
 const remote = electron.remote;
-require('../openExternal.js');
 const mist = require('../mistAPI.js');
-const syncDb = require('../syncDb.js');
 const ipcProviderWrapper = require('../ipc/ipcProviderWrapper.js');
 const BigNumber = require('bignumber.js');
 const Q = require('bluebird');
-const https = require('https');
 const Web3 = require('web3');
 const web3Admin = require('../web3Admin.js');
+const syncDb = require('../syncDb.js');
+const https = require('https');
 
+require('./include/openExternal.js');
 require('./include/setBasePath')('interface');
 
 
@@ -31,11 +31,11 @@ ipc.on('data', function(e, data) {
 // make variables globally accessable
 window.mist = mist();
 window.dirname = remote.getGlobal('dirname');
-window.syncDb = syncDb;
 window.BigNumber = BigNumber;
 window.Q = Q;
 window.web3 = new Web3(new Web3.providers.IpcProvider('', ipcProviderWrapper));
 web3Admin.extend(window.web3);
 
+window.syncDb = syncDb;
 window.ipc = ipc;
 window.https = https;
