+++ modules/preloader/wallet.js
@@ -10,8 +10,8 @@ const BigNumber = require('bignumber.js');
 const Web3 = require('web3');
 const ipcProviderWrapper = require('../ipc/ipcProviderWrapper.js');
 const web3Admin = require('../web3Admin.js');
-require('../openExternal.js');
 
+require('./include/openExternal.js');
 require('./include/setBasePath')('interface/wallet');
 
 // disable pinch zoom
