+++ modules/windows.js
@@ -26,7 +26,7 @@ class Window extends EventEmitter {
         this.isPrimary = !!opts.primary;
         this.type = type;
         this.isPopup = !!opts.isPopup;
-        this.ownerId = opts.ownerId;
+        this.ownerId = opts.ownerId; // the window which creates this new window
 
         let electronOptions = {
             title: Settings.appName,
